Moving local model parallelism example out of unit tests and into examples
